feat: refactor package structure and implement product and category features
