+++ app/src/main/java/com/proyect/ravvisant/MainActivity.kt
@@ -1,15 +1,11 @@
 package com.proyect.ravvisant
 
 import android.os.Bundle
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.WindowCompat
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
-import com.google.android.material.appbar.MaterialToolbar
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
@@ -20,27 +16,6 @@ class MainActivity : AppCompatActivity() {
         val navHostFragment = supportFragmentManager
             .findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         val navController = navHostFragment.navController
-
-        val appBarConfiguration = AppBarConfiguration(
-            setOf(
-                R.id.homeFragment,
-                R.id.searchFragment,
-                R.id.cartFragment,
-                R.id.favoriteFragment,
-                R.id.profileFragment
-            )
-        )
-
-//        val toolbar = findViewById<MaterialToolbar>(R.id.toolbar)
-//        setSupportActionBar(toolbar)
-//        setupActionBarWithNavController(navController, appBarConfiguration)
-//
-//        supportActionBar?.setDisplayShowTitleEnabled(false)
-//
-//        val toolbarTitle = toolbar.findViewById<TextView>(R.id.toolbarTitle)
-//        navController.addOnDestinationChangedListener { _, destination, _ ->
-//            toolbarTitle.text = destination.label
-//        }
 
         val bottomNav = findViewById<BottomNavigationView>(R.id.bottomNavigationView)
         bottomNav.setupWithNavController(navController)
