+++ app/src/main/java/com/proyect/ravvisant/core/utils/diff/CategoryDiffCallback.kt
@@ -1,7 +1,7 @@
-package com.proyect.ravvisant.utils.diff
+package com.proyect.ravvisant.core.utils.diff
 
 import androidx.recyclerview.widget.DiffUtil
-import com.proyect.ravvisant.data.model.Category
+import com.proyect.ravvisant.domain.model.Category
 
 class CategoryDiffCallback : DiffUtil.ItemCallback<Category>() {
     override fun areItemsTheSame(
