+++ app/src/main/java/com/proyect/ravvisant/features/auth/LoginActivity.kt
@@ -1,4 +1,4 @@
-package com.proyect.ravvisant.features.auth.presentation
+package com.proyect.ravvisant.features.auth
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,6 +7,7 @@ import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.proyect.ravvisant.MainActivity
 import com.proyect.ravvisant.R
+import com.proyect.ravvisant.features.auth.RegisterActivity
 
 class LoginActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
