+++ app/src/main/java/com/proyect/ravvisant/features/categories/CategoryViewHolder.kt
@@ -1,8 +1,7 @@
 package com.proyect.ravvisant.features.categories
 
-import androidx.lifecycle.ViewModel
 import androidx.recyclerview.widget.RecyclerView
-import com.proyect.ravvisant.data.model.Category
+import com.proyect.ravvisant.domain.model.Category
 import com.proyect.ravvisant.databinding.ItemCategoryBinding
 
 class CategoryViewHolder(private val binding: ItemCategoryBinding) :
