+++ app/src/main/java/com/proyect/ravvisant/features/categories/adapter/CategoryAdapter.kt
@@ -1,11 +1,12 @@
-package com.proyect.ravvisant.features.categories
+package com.proyect.ravvisant.features.categories.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import com.proyect.ravvisant.data.model.Category
+import com.proyect.ravvisant.core.utils.diff.CategoryDiffCallback
 import com.proyect.ravvisant.databinding.ItemCategoryBinding
-import com.proyect.ravvisant.utils.diff.CategoryDiffCallback
+import com.proyect.ravvisant.domain.model.Category
+import com.proyect.ravvisant.features.categories.CategoryViewHolder
 
 class CategoryAdapter : ListAdapter<Category, CategoryViewHolder>(CategoryDiffCallback()) {
     override fun onCreateViewHolder(
