--- app/src/main/java/com/proyect/ravvisant/features/home/presentation/HomeViewHolder.kt
@@ -1,8 +1,0 @@
-package com.proyect.ravvisant.features.home.presentation
-
-import androidx.lifecycle.ViewModel
-import androidx.recyclerview.widget.RecyclerView
-
-class HomeViewHolder : ViewModel() {
-
-}
