+++ app/src/main/java/com/proyect/ravvisant/features/product/adapters/ProductAdapter.kt
@@ -4,13 +4,30 @@ package com.proyect.ravvisant.features.product.adapters
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
 import com.proyect.ravvisant.databinding.ItemProductGridBinding
-import com.proyect.ravvisant.domain.Product
-import com.proyect.ravvisant.features.home.adapters.ProductClickCallback
-import com.proyect.ravvisant.utils.diff.ProductDiffCallback
+import com.proyect.ravvisant.domain.model.Product
+import com.proyect.ravvisant.core.common.ProductClickCallback
+import com.proyect.ravvisant.core.utils.diff.ProductDiffCallback
 
 class ProductAdapter(private val callback: ProductClickCallback) :
-    ListAdapter<Product, ProductViewHolder>(ProductDiffCallback()) {
+    ListAdapter<Product, ProductAdapter.ProductViewHolder>(ProductDiffCallback()) {
+
+    inner class ProductViewHolder(private val binding: ItemProductGridBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+        fun bind(product: Product, callback: ProductClickCallback) {
+            binding.product = product
+            binding.callback = callback
+
+            binding.root.setOnClickListener {
+                callback.onProductClick(product)
+            }
+
+            binding.executePendingBindings()
+        }
+    }
+
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int
