+++ app/src/main/java/com/proyect/ravvisant/features/product/viewmodel/ProductDetailViewModel.kt
@@ -1,9 +1,9 @@
-package com.proyect.ravvisant.features.product.adapters
+package com.proyect.ravvisant.features.product.viewmodel
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.proyect.ravvisant.domain.Product
+import com.proyect.ravvisant.domain.model.Product
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
