+++ app/src/main/res/layout/fragment_product_detail.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="product"
-            type="com.proyect.ravvisant.domain.Product" />
+            type="com.proyect.ravvisant.domain.model.Product" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
